Attempt at fixing Travis
Looks like the package dependency for cypher-shell is not reliably defined for all Ubuntu distributions, so why not try to install it manually if available. The declared dependency may also be different for the patch versions of the neo4j package. I guess it's generally useful to work with the most recent version.