Avoid use of experimental bytes_to_utf8
Not too sure about this one ... while bytes_to_utf8 is "experimental" according to perlapi, it isn't labelled as such in perlguts, which actually recommends using it for our exact scenario. bytes_to_utf8 was introduced in Perl 5.7.0, so it's been around like forever and it seems unlikely that it would suddenly vanish. And avoiding bytes_to_utf8 in the way this change does uses up more (temp) memory.

In spite of all that, using "experimental" functions seems like a bad idea to me.

newSVpvn_flags is available from Perl 5.10.1.