Test more Bolt versions, fail on connection error
I don't think Neo4j::Bolt necessarily has to be able to speak all these Bolt 5.x variants. Just a single Bolt 5.x variant would probably be fine, as Neo4j 5.x servers apparently understand all of them. But for development, it's useful to know exactly where we stand.

https://neo4j.com/docs/bolt/current/bolt-compatibility/

The motivation for this change is #54. As noted in that report, at the time Neo4j::Client only declared Bolt 5.2+ as supported (Neo4j 5.7+). That is still the case today. Tests on any earlier versions are expected to fall back to Bolt 4.0.

Some Neo4j versions refuse the connection entirely, leading the crucial tests to be skipped and giving a false positive result for the test suite. This change introduces a new test to address that. The new test is only run when live DB tests have been explicitly requested by the user.