Add Neo4j::Types v2 spatial/temporal support
The earlier code for Neo4j::Bolt struct types is now functionally redundant, but remains in place as a special case because it's faster by an order of magnitude (~ 7x). This difference in performance is only natural, since the flexibility provided by Neo4j::Types comes at the cost of calling a bunch of Perl methods.

The perlapi call sv_isa_sv() could speed this up some, but is currently still experimental.

The special case is bracketed with #ifdef directives, which allow for disabling it if desired. The code works either way, this is just about speed vs. the need to maintain dual codepaths.

Relevant tests are added. All types tests are gathered using numbers in the range 020..028 for consistency.