Refactor SV_to_neo4j_value()
Moves the "is object" check further to the outside in the condition tree, so that it becomes easier to add additional object checks, for example checks for Neo4j::Types objects other than those that are also Neo4j::Bolt objects.

Also changes the scalar condition tree to allow the stable number/string tracking introduced in Perl 5.36 to work. (#42)

For rare true dualvars (variables specially crafted to have independent number/string parts, such as $!), the new code now selects the string part instead of the number part. There is no way for Neo4j::Bolt to actually know which part the user wants, but I suspect that the string part at least has the potential to yield a more useful error message, which could aid debugging.

Stable boolean tracking and Perl core bool support are not added with this change.