Add warning for outdated libneo4j-omni versions
Neo4j::Bolt depends on libneo4j-client/omni features that aren't available in all versions. Using this module with a library version which is too old may fail. Such failures aren't always obvious. A library version check and Perl warning may help users to keep their software current.

The Neo4j::Client prereq version is already declared in Makefile.PL, but this is only relevant for CPAN. Users performing a manual installation with a tarball or from GitHub may be oblivious to the fact that their library version could be outdated. This is a situation we have seen.

It's not necessary to have this warning in a BEGIN block, but keeping it close to the module version declaration and the XS loading seems sensible.