Fix memory leaks with error messages and protocol version
* Server evaluation errors (failure details) in Neo4j::Bolt::ResultStream are affected by memory leaks. Buffers are allocated and never freed.

Care must be taken to free the error message memory not only in the destructor, but also before assigning another error message to that same pointer.

* Client error messages in Neo4j::Bolt::Cxn, Neo4j::Bolt::Txn, and Neo4j::Bolt::ResultStream objects are affected by memory leaks. In each case, a buffer is allocated and never freed.

The string pointer stored in the object usually doesn't actually point to the buffer, but to a string literal that cannot be freed. A solution is to copy the error message into a second buffer, which is something that savepv() makes very easy to do. That second buffer can then be reliably freed in the destructor, and the first buffer can be converted into a local char[] array that doesn't need to be freed later.

* Calling protocol_version_() usually allocates a string buffer for the version number and never frees it. However, it may also return a string literal that cannot be freed.

The solution is to create the buffer using Perl_form(), which will automatically free it once the XS logic has copied its contents into the SV upon return. Another solution would be to declare the XSUB with a CLEANUP: Safefree(RETVAL) section and savepvs() the string literal.